Rename of project to ensembl_lite and command line tool to elt
