+++ src/ensembl_lite/_aligndb.py
@@ -2,7 +2,7 @@ import typing
 
 import numpy
 
-from ensembl_cli._db_base import SqliteDbMixin
+from ensembl_lite._db_base import SqliteDbMixin
 
 
 class AlignRecordType(typing.TypedDict):
