+++ src/ensembl_lite/_genomedb.py
@@ -3,7 +3,7 @@ import typing
 from cogent3 import get_app
 from cogent3.app.composable import define_app
 
-from ensembl_cli._db_base import SqliteDbMixin
+from ensembl_lite._db_base import SqliteDbMixin
 
 
 OptInt = typing.Optional[int]
