+++ src/ensembl_lite/data/sample.cfg
