+++ src/ensembl_lite/download.py
@@ -9,11 +9,11 @@ import click
 
 from cogent3 import load_tree
 
-from ensembl_cli._config import Config
-from ensembl_cli._site_map import get_site_map
-from ensembl_cli.ftp_download import download_data, listdir
-from ensembl_cli.species import Species, species_from_ensembl_tree
-from ensembl_cli.util import (
+from ensembl_lite._config import Config
+from ensembl_lite._site_map import get_site_map
+from ensembl_lite.ftp_download import download_data, listdir
+from ensembl_lite.species import Species, species_from_ensembl_tree
+from ensembl_lite.util import (
     dont_checksum,
     get_resource_path,
     is_signature,
